Cleaned up octree

+++ Assets/Scripts/Utils/Noise/Noise2D.cs
@@ -3,5 +3,5 @@ using System.Collections;
 
 public abstract class Noise2D {
 
-    public abstract int generate(int x, int y);
+    public abstract int generate(float x, float y);
 }
